testsuites/psxtests: Correct comparisons of different signedness
The GCC warning -Wsign-compare flagged these instances of comparing
signed and unsigned types. A common pair was size_t and int.

+++ testsuites/psxtests/psxfile01/test_cat.c
@@ -50,7 +50,7 @@
 #include <tmacros.h>
 
 /* forward declarations to avoid warnings */
-void test_cat(char *file, int offset_arg, int length);
+void test_cat(char *file, int offset_arg, size_t length);
 
 /*
  *  test_cat routine
@@ -59,15 +59,15 @@ void test_cat(char *file, int offset_arg, int length);
 unsigned char test_cat_buffer[ 1024 ];
 
 void test_cat(
-  char *file,
-  int   offset_arg,
-  int   length
+  char   *file,
+  int     offset_arg,
+  size_t  length
 )
 {
   int            fd;
   int            status;
   int            is_printable = 0;
-  int            my_length;
+  size_t         my_length;
   int            i;
   unsigned char  c;
   int            count = 0;
